+++ cocos2d/cocos/2d/CCGrid.cpp
@@ -430,7 +430,7 @@ void Grid3D::calculateVertexPoints()
     {
         for (y = 0; y < _gridSize.height; ++y)
         {
-            int idx = (int)(y * _gridSize.width) + x;
+            int idx = static_cast<int>(y * _gridSize.width + x);
 
             float x1 = x * _step.x + _gridRect.origin.x;
             float x2 = x1 + _step.x;
@@ -794,7 +794,7 @@ void TiledGrid3D::updateVertexBuffer()
     }
     _drawCommand.updateVertexBuffer(_vertexBuffer, (unsigned int)(numOfPoints * sizeof(Vec3) + numOfPoints * sizeof(Vec2)) );
 
-    _drawCommand.updateIndexBuffer(_indices, gradSize * 6 * sizeof(unsigned short));
+    _drawCommand.updateIndexBuffer(_indices, static_cast<int>(gradSize * 6 * sizeof(unsigned short)));
 }
 
 void TiledGrid3D::updateVertexAndTexCoordinate()
@@ -809,11 +809,11 @@ void TiledGrid3D::updateVertexAndTexCoordinate()
         memcpy((char*)_vertexBuffer + offset, &tempVecPointer[i], sizeof(Vec3));
         memcpy((char*)_vertexBuffer + offset + sizeof(Vec3), &tempTexPointer[i], sizeof(Vec2));
     }
-    _drawCommand.createVertexBuffer((unsigned int)(sizeof(Vec3) + sizeof(Vec2) ), numOfPoints, CustomCommand::BufferUsage::DYNAMIC);
-    _drawCommand.updateVertexBuffer(_vertexBuffer, numOfPoints * sizeof(Vec3) + numOfPoints * sizeof(Vec2));
+    _drawCommand.createVertexBuffer((unsigned int)(sizeof(Vec3) + sizeof(Vec2) ), static_cast<int>(numOfPoints), CustomCommand::BufferUsage::DYNAMIC);
+    _drawCommand.updateVertexBuffer(_vertexBuffer, static_cast<int>(numOfPoints * sizeof(Vec3) + numOfPoints * sizeof(Vec2)));
 
-    _drawCommand.createIndexBuffer(CustomCommand::IndexFormat::U_SHORT, gradSize * 6, CustomCommand::BufferUsage::DYNAMIC);
-    _drawCommand.updateIndexBuffer(_indices, gradSize * 6 * sizeof(unsigned short));
+    _drawCommand.createIndexBuffer(CustomCommand::IndexFormat::U_SHORT, static_cast<int>(gradSize * 6), CustomCommand::BufferUsage::DYNAMIC);
+    _drawCommand.updateIndexBuffer(_indices, static_cast<int>(gradSize * 6 * sizeof(unsigned short)));
 }
 
 NS_CC_END
